BRD-MasterAdmin-Prudhvi/BRD-MasterAdmin-Prudhvi/reporting/serializers.py
from rest_framework import serializers
from .models import Report, Analytics

class ReportSerializer(serializers.ModelSerializer):
    class Meta:
        model = Report
        fields = '__all__'

class AnalyticsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Analytics
        fields = '__all__'
